refactor: move agl requests to separate folders
