+++ src/pages/SignUpPage/SignUpPage.tsx
@@ -7,9 +7,9 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import { useMutation } from '@apollo/client';
 import { RoutePath } from '../../constants/routeVariables';
 import { schema } from '../../utils/validationSchema';
-import { SIGNUP } from '../../graphql/authentication/mutation';
 import { authService } from '../../graphql/authentication/authService';
 import { ISignupResult } from '../../graphql/authentication/authResult.interface.';
+import { SIGNUP } from '../../graphql/mutations/auth';
 import { IFormInput } from './formInput.interface';
 import {
   ButtonLink,
